fix: couple of off by one errors
The index for book id -> book name was off by one. Also the
NT SEDRA book ids were off by one to compensate. This meant
modules started with Gen and missed out Revelation (65 books)
but labeled the books Exo - Rev.

With this commit we now have the full 66 books and they seem to
be labeled correctly.

Signed-off-by: David James McCorrie <djmccorrie@gmail.com>