feat: strip off prefix and suffix
Getting closer to doing some proper regular verb morphology evaluation.
The unique word count total is now down to 29k, due to pulling off some
prefixes and suffixes.

This implementation is still quite naive, which means there are some
false positives. The next step is to use the prefix and suffix to
determine the tense/mood/number/gender/stem.

The other issue is that sometimes prepositions are combined in full.
This code only strips off one of them and treats the second as part of
the word itself.

Signed-off-by: David James McCorrie <djmccorrie@gmail.com>